requirements.txt
aiohttp==3.6.3
altgraph==0.16.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord.py==1.5.1
future==0.18.1
idna==2.10
multidict==4.7.6
pefile==2019.4.18
py2exe==0.9.2.2
pyinstaller @ https://github.com/pyinstaller/pyinstaller/archive/develop.tar.gz
pyinstaller-hooks-contrib==2020.11
python-dotenv==0.15.0
pywin32-ctypes==0.2.0
requests==2.25.1
urllib3==1.26.2
yarl==1.5.1
